requirements.txt
Django>=4.0.1,<=4.1.5
djangorestframework>=3.12.4,!=3.14.0
sqlparse==0.4.2
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.22.1,<0.23
django-multiselectfield
